requirements.txt
absl-py==2.0.0
accelerate==0.21.0
aiohttp==3.8.6
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.3
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.2.post2
bleach==6.0.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.0
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.7
cycler==0.11.0
datasets==2.15.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
dnspython==2.7.0
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.7.0
email_validator==2.2.0
evaluate==0.4.1
executing==1.2.0
fastapi==0.105.0
fastjsonschema==2.18.0
filelock==3.12.2
firebase-admin==6.3.0
fonttools==4.41.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
grpclib==0.4.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.19.4
hyperframe==6.0.1
idna==3.4
importlib-metadata==7.0.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.23.0
kiwisolver==1.4.4
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
modal==0.56.4287
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
nbclient==0.8.0
nbconvert==7.7.2
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==7.0.0
notebook_shim==0.2.3
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py3==7.352.0
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
orjson==3.10.13
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
peft==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.10.4
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rouge-score==0.1.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.1
scipy==1.11.4
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.37.1
setproctitle==1.3.3
shtab==1.6.4
sigtools==4.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
synchronicity==0.5.3
tblib==3.0.0
tensorboardX==2.6.2.2
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
triton==2.0.0
trl==0.7.4
typer==0.9.0
types-certifi==2021.10.8.3
types-toml==0.10.8.7
typing_extensions==4.12.2
tyro==0.5.18
tzdata==2023.3
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.34.0
uvloop==0.21.0
wandb==0.16.0
watchfiles==0.21.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==14.1
widgetsnbextension==4.0.8
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
